# CLN: replace lambdas with named functions so they are labeled in asv